+++ app/components/frost-bunsen-input-hidden.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-frost-bunsen/components/inputs/hidden'
+export {default} from 'ember-frost-bunsen/components/inputs/hidden'