+++ config/coverage.js
@@ -1,13 +1,10 @@
 module.exports = {
   coverageEnvVar: 'COVERAGE',
   coverageFolder: 'coverage',
-  excludes: [
-    '**/mirage/**/*'
-  ],
-  useBabelInstrumenter: true,
   reporters: [
     'html',
     'lcov',
     'text-summary'
-  ]
+  ],
+  useBabelInstrumenter: true
 }