+++ tests/dummy/app/pods/model/formats/models/ipv6-prefix.js
@@ -1,0 +1,9 @@
+export default {
+  properties: {
+    foo: {
+      format: 'ipv6-prefix',
+      type: 'string'
+    }
+  },
+  type: 'object'
+}